Use force delete over delete.
Openstack has a featured labelled "soft-delete", it emails cloud
admins to restore a VM for a certain amount of time before a VM
is fully removed.

It functions very differently to a standard delete, for a start,
the VM is in a state of "soft-deleted" and all volumes, VIPs,
security groups, etc. associated with the VM are still seen to
be in use.

This causes a great mess for terraform, even if you modify
the code to see a "soft-deleted" state you would have to do
some polling for who knows how long to delete the instance
and associated items.

By using ForceDelete we bypass the soft-delete functionality
if enabled, thus terraform operates as expected.

REF: https://github.com/hashicorp/terraform/issues/5104

Signed-off-by: Ian Duffy <ian@ianduffy.ie>